requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
sqlalchemy==2.0.23
pydantic==2.9.2
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-jose==3.3.0
alembic==1.13.1
